chore: rename `run` to `setup`
