+++ cmd/relayer/relayer.go
@@ -1,7 +1,7 @@
 package relayer
 
 import (
-	"github.com/dymensionxyz/roller/cmd/relayer/run"
+	"github.com/dymensionxyz/roller/cmd/relayer/setup"
 	"github.com/dymensionxyz/roller/cmd/relayer/start"
 	"github.com/dymensionxyz/roller/cmd/relayer/status"
 	"github.com/dymensionxyz/roller/cmd/services"
@@ -18,7 +18,7 @@ func Cmd() *cobra.Command {
 		Short: "Commands for running and managing the RollApp relayer.",
 	}
 
-	cmd.AddCommand(run.Cmd())
+	cmd.AddCommand(setup.Cmd())
 	cmd.AddCommand(start.Cmd())
 	cmd.AddCommand(status.Cmd())
