+++ cmd/relayer/setup/setup.go
@@ -1,4 +1,4 @@
-package run
+package setup
 
 import (
 	"encoding/json"
